+++ src/storage/stored_value.rs
@@ -1,5 +1,5 @@
 use crate::storage::{in_memory::StoreError, value::Value};
-use std::collections::VecDeque;
+use std::collections::{HashSet, VecDeque};
 use std::time::Instant;
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -19,14 +19,14 @@ impl StoredValue {
     pub(crate) fn value(&self) -> Result<&str, StoreError> {
         match &self.value {
             Value::String(value) => Ok(value),
-            Value::List(_) => Err(StoreError::WrongType),
+            Value::List(_) | Value::Set(_) => Err(StoreError::WrongType),
         }
     }
 
     pub(crate) fn value_mut(&mut self) -> Result<&mut String, StoreError> {
         match &mut self.value {
             Value::String(value) => Ok(value),
-            Value::List(_) => Err(StoreError::WrongType),
+            Value::List(_) | Value::Set(_) => Err(StoreError::WrongType),
         }
     }
 
@@ -37,7 +37,7 @@ impl StoredValue {
     pub(crate) fn into_value(self) -> Result<String, StoreError> {
         match self.value {
             Value::String(value) => Ok(value),
-            Value::List(_) => Err(StoreError::WrongType),
+            Value::List(_) | Value::Set(_) => Err(StoreError::WrongType),
         }
     }
     pub(crate) fn expires_at(&self) -> Option<Instant> {
@@ -73,14 +73,35 @@ impl StoredValue {
     pub(crate) fn list(&self) -> Result<&VecDeque<String>, StoreError> {
         match &self.value {
             Value::List(values) => Ok(values),
-            Value::String(_) => Err(StoreError::WrongType),
+            Value::String(_) | Value::Set(_) => Err(StoreError::WrongType),
         }
     }
 
     pub(crate) fn list_mut(&mut self) -> Result<&mut VecDeque<String>, StoreError> {
         match &mut self.value {
             Value::List(values) => Ok(values),
-            Value::String(_) => Err(StoreError::WrongType),
+            Value::String(_) | Value::Set(_) => Err(StoreError::WrongType),
+        }
+    }
+
+    pub(crate) fn new_set() -> Self {
+        Self {
+            value: Value::Set(HashSet::new()),
+            expires_at: None,
+        }
+    }
+
+    pub(crate) fn set(&self) -> Result<&HashSet<String>, StoreError> {
+        match &self.value {
+            Value::Set(values) => Ok(values),
+            Value::String(_) | Value::List(_) => Err(StoreError::WrongType),
+        }
+    }
+
+    pub(crate) fn set_mut(&mut self) -> Result<&mut HashSet<String>, StoreError> {
+        match &mut self.value {
+            Value::Set(values) => Ok(values),
+            Value::String(_) | Value::List(_) => Err(StoreError::WrongType),
         }
     }
 }