+++ src/storage/tests/mod.rs
@@ -2,6 +2,7 @@ mod expiration;
 mod keys;
 mod lists;
 mod numeric;
+mod sets;
 mod stored_value;
 mod strings;
 mod wrong_type;