+++ src/storage/value.rs
@@ -1,7 +1,8 @@
-use std::collections::VecDeque;
+use std::collections::{HashSet, VecDeque};
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub(crate) enum Value {
     String(String),
     List(VecDeque<String>),
+    Set(HashSet<String>),
 }